# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /updates/{packageSlug}.xml:
    get:
      operationId: getPackageUpdates
      tags: [updates]
      summary: License-gated Joomla update manifest
      description: |
        Returns the Joomla `<updates>` XML manifest for the given package
        slug. `pkg_aiboost` (free) always returns the latest release.
        `pkg_aiboost_pro` requires a valid `key` whose `domain` matches one
        of the activated domains on the license; anything else returns an
        empty but valid `<updates/>` document so Joomla silently shows no
        updates available.
      parameters:
        - in: path
          name: packageSlug
          required: true
          schema:
            type: string
            enum: [pkg_aiboost, pkg_aiboost_pro]
      x-query-parameters: |
        Joomla appends these via the onInstallerBeforeFetchManifest hook:
          key    — license key (required for pkg_aiboost_pro)
          domain — site hostname (required for pkg_aiboost_pro)
          v      — currently installed version (analytics only)
        They are intentionally omitted from the spec to keep the
        generated typed client free of name collisions; the server
        reads them directly from req.query.
      responses:
        "200":
          description: Joomla update manifest XML (possibly empty)
          content:
            application/xml:
              schema:
                type: string
        "404":
          description: Unknown package slug
          content:
            application/xml:
              schema:
                type: string
...